requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.8.1=pypi_0
astor=0.8.0=pypi_0
beautifulsoup4=4.9.3=pypi_0
certifi=2019.9.11=pypi_0
chardet=3.0.4=pypi_0
click=7.0=pypi_0
flask=1.1.1=pypi_0
gast=0.3.2=pypi_0
grpcio=1.25.0=pypi_0
gtts=2.2.2=pypi_0
gtts-token=1.1.4=pypi_0
h5py=2.10.0=pypi_0
idna=2.8=pypi_0
itsdangerous=1.1.0=pypi_0
jinja2=2.10.3=pypi_0
joblib=0.14.0=pypi_0
keras=2.2.4=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
markdown=3.1.1=pypi_0
markupsafe=1.1.1=pypi_0
numpy=1.17.4=pypi_0
pillow=6.2.1=pypi_0
pip=21.0.1=py36haa95532_0
playsound=1.2.2=pypi_0
protobuf=3.10.0=pypi_0
python=3.6.13=h3758d61_0
python_abi=3.6=1_cp36m
pyyaml=5.1.2=pypi_0
requests=2.22.0=pypi_0
scikit-learn=0.21.3=pypi_0
scipy=1.0.0=pypi_0
setuptools=39.1.0=pypi_0
six=1.13.0=pypi_0
sklearn=0.0=pypi_0
soupsieve=1.9.5=pypi_0
sqlite=3.35.4=h2bbff1b_0
tensorboard=1.9.0=pypi_0
tensorflow=1.9.0=pypi_0
termcolor=1.1.0=pypi_0
urllib3=1.25.7=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=0.16.0=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py36h7fe50ca_0
